Split the MessageEditor service and factory
