+++ twake/frontend/src/app/scenes/Apps/Messages/Parts/Thread.tsx
@@ -3,7 +3,7 @@ import './Threads.scss';
 import Draggable from 'components/Draggable/Draggable.js';
 import UploadZone from 'components/Uploads/UploadZone.js';
 import Workspaces from 'services/workspaces/workspaces.js';
-import MessageEditorsManager from 'app/services/Apps/Messages/MessageEditors';
+import MessageEditorsManager from 'app/services/Apps/Messages/MessageEditorServiceFactory';
 import { Message } from 'app/services/Apps/Messages/Message';
 
 type Props = {
