+++ twake/frontend/src/app/services/Apps/Messages/MessageEditorServiceFactory.ts
@@ -1,0 +1,22 @@
+import { MessageEditorService } from "./MessageEditorService";
+
+class MessageEditorServiceFactory {
+  private services = new Map<string, MessageEditorService>();
+
+  get(channelId: string): MessageEditorService {
+    let service = this.services.get(channelId);
+
+    if (service) {
+      return service;
+    }
+    
+    service = new MessageEditorService(channelId);
+    this.services.set(channelId, service);
+
+    return service;
+  }
+}
+
+const factory = new MessageEditorServiceFactory();
+(window as any).TwakeMessageEditorServiceFactory = factory;
+export default factory;
